Continue test file restructuring for faster incremental compilation
Split additional test suites from test_integration.cc into focused files:
- test_integration_errors.cc: error_path_tests, network_error_tests
- test_integration_reactor.cc: reactor_error_tests, reactor_mask_tests
- test_integration_advanced.cc: interceptor_tests, server_config_tests,
  defensive_sync_tests

Fixed port conflicts by assigning unique port offsets (160+) to avoid
test isolation issues with other test files.

This completes the restructuring started in 9610ad8, reducing the main
test file from 4,097 to 2,323 lines (43% reduction). All 144 tests pass.
