+++ tests/test_integration.cc
@@ -49,7 +49,7 @@ BOOST_AUTO_TEST_SUITE(additional_coverage_tests)
 
 BOOST_FIXTURE_TEST_CASE(max_request_size_enforcement, IntegrationFixture)
 {
-  start_server(1, 60);
+  start_server(1, 160);
   server().set_max_request_sz(500);  // Very small limit
 
   auto client = create_client();
@@ -61,7 +61,7 @@ BOOST_FIXTURE_TEST_CASE(max_request_size_enforcement, IntegrationFixture)
 
 BOOST_FIXTURE_TEST_CASE(unknown_method_error, IntegrationFixture)
 {
-  start_server(1, 61);
+  start_server(1, 161);
   auto client = create_client();
 
   Response r = client->execute("nonexistent_method", Value("test"));
@@ -72,7 +72,7 @@ BOOST_FIXTURE_TEST_CASE(unknown_method_error, IntegrationFixture)
 BOOST_FIXTURE_TEST_CASE(server_log_message, IntegrationFixture)
 {
   std::ostringstream log_stream;
-  start_server(1, 62);
+  start_server(1, 162);
   server().log_errors(&log_stream);
 
   auto client = create_client();
@@ -84,7 +84,7 @@ BOOST_FIXTURE_TEST_CASE(server_log_message, IntegrationFixture)
 
 BOOST_FIXTURE_TEST_CASE(multiple_dispatchers, IntegrationFixture)
 {
-  start_server(1, 63);
+  start_server(1, 163);
   server().enable_introspection();
 
   auto client = create_client();
@@ -272,185 +272,9 @@ BOOST_AUTO_TEST_SUITE_END()
 
 // firewall_tests moved to test_integration_security.cc
 
-//=============================================================================
-// Interceptor Tests
-//=============================================================================
-BOOST_AUTO_TEST_SUITE(interceptor_tests)
-
-// Simple interceptor that counts calls
-class CountingInterceptor : public Interceptor {
-public:
-  static int call_count;
-
-  void process(Method* m, const Param_list& params, Value& result) override {
-    ++call_count;
-    yield(m, params, result);
-  }
-};
-
-int CountingInterceptor::call_count = 0;
-
-BOOST_FIXTURE_TEST_CASE(interceptor_invoked, IntegrationFixture)
-{
-  start_server(1, 85);
-  CountingInterceptor::call_count = 0;
-  server().push_interceptor(new CountingInterceptor());
-
-  auto client = create_client();
-  Response r1 = client->execute("echo", Value("test1"));
-  Response r2 = client->execute("echo", Value("test2"));
-
-  BOOST_CHECK(!r1.is_fault());
-  BOOST_CHECK(!r2.is_fault());
-  BOOST_CHECK_EQUAL(CountingInterceptor::call_count, 2);
-}
-
-// Interceptor that modifies the result
-class ModifyingInterceptor : public Interceptor {
-public:
-  void process(Method* m, const Param_list& params, Value& result) override {
-    yield(m, params, result);
-    // Append suffix to string results
-    if (result.is_string()) {
-      result = Value(result.get_string() + "_modified");
-    }
-  }
-};
-
-BOOST_FIXTURE_TEST_CASE(interceptor_modifies_result, IntegrationFixture)
-{
-  start_server(1, 86);
-  server().push_interceptor(new ModifyingInterceptor());
-
-  auto client = create_client();
-  Response r = client->execute("echo", Value("test"));
-  BOOST_CHECK(!r.is_fault());
-  BOOST_CHECK_EQUAL(r.value().get_string(), "test_modified");
-}
-
-// Chained interceptors
-BOOST_FIXTURE_TEST_CASE(chained_interceptors, IntegrationFixture)
-{
-  start_server(1, 87);
-  CountingInterceptor::call_count = 0;
-  server().push_interceptor(new CountingInterceptor());
-  server().push_interceptor(new CountingInterceptor());
-
-  auto client = create_client();
-  Response r = client->execute("echo", Value("test"));
-  BOOST_CHECK(!r.is_fault());
-  BOOST_CHECK_EQUAL(CountingInterceptor::call_count, 2);  // Both interceptors called
-}
-
-BOOST_AUTO_TEST_SUITE_END()
-
-//=============================================================================
-// Error Path Tests
-//=============================================================================
-BOOST_AUTO_TEST_SUITE(error_path_tests)
-
-BOOST_FIXTURE_TEST_CASE(method_throws_fault, IntegrationFixture)
-{
-  start_server(1, 90);
-  auto client = create_client();
-
-  Response r = client->execute("error_method", Value(""));
-  BOOST_CHECK(r.is_fault());
-  BOOST_CHECK_EQUAL(r.fault_code(), 123);
-}
-
-BOOST_FIXTURE_TEST_CASE(method_returns_complex_fault, IntegrationFixture)
-{
-  start_server(1, 91);
-  auto client = create_client();
-
-  // error_method throws Fault(123, "My fault")
-  Response r = client->execute("error_method", Value("test"));
-  BOOST_CHECK(r.is_fault());
-  std::string fault_str = r.fault_string();
-  BOOST_CHECK(fault_str.find("fault") != std::string::npos ||
-              fault_str.find("Fault") != std::string::npos);
-}
-
-BOOST_FIXTURE_TEST_CASE(sequential_errors, IntegrationFixture)
-{
-  start_server(1, 92);
-  auto client = create_client();
-
-  // Multiple error calls should all work
-  for (int i = 0; i < 3; ++i) {
-    Response r = client->execute("error_method", Value(i));
-    BOOST_CHECK(r.is_fault());
-  }
-}
-
-BOOST_FIXTURE_TEST_CASE(error_then_success, IntegrationFixture)
-{
-  start_server(1, 93);
-  auto client = create_client();
-
-  Response r1 = client->execute("error_method", Value(""));
-  BOOST_CHECK(r1.is_fault());
-
-  // Server should still work after error
-  Response r2 = client->execute("echo", Value("still working"));
-  BOOST_CHECK(!r2.is_fault());
-  BOOST_CHECK_EQUAL(r2.value().get_string(), "still working");
-}
-
-BOOST_AUTO_TEST_SUITE_END()
-
-//=============================================================================
-// Server Configuration Tests
-//=============================================================================
-BOOST_AUTO_TEST_SUITE(server_config_tests)
-
-BOOST_FIXTURE_TEST_CASE(verification_level_weak, IntegrationFixture)
-{
-  start_server(1, 95);
-  server().set_verification_level(http::HTTP_CHECK_WEAK);
-
-  auto client = create_client();
-  Response r = client->execute("echo", Value("weak"));
-  BOOST_CHECK(!r.is_fault());
-}
-
-BOOST_FIXTURE_TEST_CASE(verification_level_strict, IntegrationFixture)
-{
-  start_server(1, 96);
-  server().set_verification_level(http::HTTP_CHECK_STRICT);
-
-  auto client = create_client();
-  Response r = client->execute("echo", Value("strict"));
-  BOOST_CHECK(!r.is_fault());
-}
-
-BOOST_FIXTURE_TEST_CASE(get_reactor, IntegrationFixture)
-{
-  start_server(1, 97);
-
-  // Verify reactor is accessible
-  BOOST_CHECK(server().get_reactor() != nullptr);
-
-  auto client = create_client();
-  Response r = client->execute("echo", Value("reactor_test"));
-  BOOST_CHECK(!r.is_fault());
-}
-
-BOOST_FIXTURE_TEST_CASE(max_request_size_boundary, IntegrationFixture)
-{
-  start_server(1, 98);
-  server().set_max_request_sz(10000);  // 10KB limit
-
-  auto client = create_client();
-
-  // Just under the limit should work
-  std::string data(500, 'x');
-  Response r = client->execute("echo", Value(data));
-  BOOST_CHECK(!r.is_fault());
-}
-
-BOOST_AUTO_TEST_SUITE_END()
+// NOTE: interceptor_tests moved to test_integration_advanced.cc
+// NOTE: error_path_tests moved to test_integration_errors.cc
+// NOTE: server_config_tests moved to test_integration_advanced.cc
 
 // connection_reuse_tests moved to test_integration_connection.cc
 
@@ -693,7 +517,7 @@ BOOST_FIXTURE_TEST_CASE(request_too_large_returns_error, IntegrationFixture)
 // Covers HTTP error path for missing Content-Length
 BOOST_FIXTURE_TEST_CASE(missing_content_length_returns_error, IntegrationFixture)
 {
-    start_server(1, 116);
+    start_server(1, 117);
 
     std::string request =
         "POST /RPC2 HTTP/1.1\r\n"
@@ -712,7 +536,7 @@ BOOST_FIXTURE_TEST_CASE(missing_content_length_returns_error, IntegrationFixture
 // Test unsupported content type in strict mode
 BOOST_FIXTURE_TEST_CASE(unsupported_content_type_strict_mode, IntegrationFixture)
 {
-    start_server(1, 117);
+    start_server(1, 118);
     server().set_verification_level(http::HTTP_CHECK_STRICT);
 
     std::string request =
@@ -734,582 +558,12 @@ BOOST_AUTO_TEST_SUITE_END()
 
 // ssl_tests moved to test_integration_ssl.cc
 
-//=============================================================================
-// Network Error Path Tests - covers connector.cc, http_client.cc error paths
-//
-// These tests exercise error handling in network operations:
-//   - Client timeout (connector.cc lines 64-66)
-//   - Connection refused (connector.cc lines 35-37)
-//   - HTTP client timeout (http_client.cc line 34)
-//   - Connection closed by peer (http_client.cc lines 61-62)
-//   - SSL certificate/key errors (ssl_lib.cc lines 213-218)
-//=============================================================================
+// NOTE: network_error_tests moved to test_integration_errors.cc
 
-BOOST_AUTO_TEST_SUITE(network_error_tests)
+// NOTE: reactor_error_tests moved to test_integration_reactor.cc
+// NOTE: reactor_mask_tests moved to test_integration_reactor.cc
 
-// Test client connection timeout
-// Covers connector.cc lines 64-66: Client_timeout exception
-BOOST_AUTO_TEST_CASE(client_connection_timeout)
-{
-    // Connect to a non-routable IP address to trigger timeout
-    // Using TEST-NET-1 (192.0.2.0/24) which is reserved for documentation
-    // and should not route, causing connection to hang and timeout
-    Inet_addr addr("192.0.2.1", 12345);
-
-    try {
-        Client<Http_client_connection> client(addr);
-        client.set_timeout(1);  // 1 second timeout
-
-        // This should timeout since the address is not routable
-        client.execute("echo", Value("timeout test"));
-
-        // If we get here without exception, the test is inconclusive
-        // (network might have routed the connection somehow)
-        BOOST_TEST_MESSAGE("Connection did not timeout - test inconclusive");
-    } catch (const Client_timeout&) {
-        // Expected - timeout occurred
-        BOOST_CHECK(true);
-    } catch (const iqnet::network_error&) {
-        // Also acceptable - immediate connection failure
-        BOOST_CHECK(true);
-    }
-}
-
-// Test connection refused error
-// Covers connector.cc lines 35-37: network_error in handle_output
-BOOST_AUTO_TEST_CASE(client_connection_refused)
-{
-    // Connect to localhost on a port that should not have a listener
-    // Port 1 is reserved and very unlikely to have a service
-    Inet_addr addr("127.0.0.1", 59998);
-
-    try {
-        Client<Http_client_connection> client(addr);
-        client.set_timeout(5);
-
-        // This should fail with connection refused
-        client.execute("echo", Value("refused test"));
-
-        BOOST_FAIL("Expected network_error for connection refused");
-    } catch (const iqnet::network_error&) {
-        // Expected - connection refused
-        BOOST_CHECK(true);
-    } catch (const Client_timeout&) {
-        // Also acceptable - might timeout instead of immediate refuse
-        BOOST_CHECK(true);
-    }
-}
-
-// Test HTTP client timeout exception type
-// Covers http_client.cc line 34: Client_timeout exception path
-// Note: We don't actually trigger a timeout here to avoid server memory leaks
-// from abruptly terminated connections. Instead we verify the timeout mechanism
-// works by testing with a normal request and checking timeout can be set.
-BOOST_FIXTURE_TEST_CASE(http_client_request_timeout, IntegrationFixture)
-{
-    start_server(1, 120);
-
-    // Create client with reasonable timeout
-    std::unique_ptr<Client_base> client(
-        new Client<Http_client_connection>(Inet_addr("127.0.0.1", port_)));
-
-    // Verify timeout can be set (covers timeout option handling)
-    client->set_timeout(30);
-
-    // Execute a normal request to ensure connection works
-    Response r = client->execute("echo", Value("timeout test"));
-    BOOST_CHECK(!r.is_fault());
-    BOOST_CHECK_EQUAL(r.value().get_string(), "timeout test");
-
-    // Verify Client_timeout exception type exists and works
-    BOOST_CHECK_NO_THROW({
-        try {
-            throw Client_timeout();
-        } catch (const std::exception& e) {
-            BOOST_CHECK(e.what() != nullptr);
-        }
-    });
-}
-
-// Test connection to server that immediately closes
-// This indirectly covers http_client.cc lines 61-62: Connection closed by peer
-BOOST_AUTO_TEST_CASE(connection_closed_by_peer)
-{
-    // Create a server socket that accepts but immediately closes
-    Socket server_sock;
-    Inet_addr addr("127.0.0.1", 0);  // Let OS choose port
-    server_sock.bind(addr);
-    server_sock.listen(1);
-    Inet_addr server_addr = server_sock.get_addr();
-
-    // Start a thread that accepts and immediately closes
-    std::thread acceptor([&server_sock]() {
-        try {
-            Socket accepted = server_sock.accept();
-            // Send partial HTTP response then close
-            const char* partial = "HTTP/1.1 200";
-            accepted.send(partial, strlen(partial));
-            std::this_thread::sleep_for(std::chrono::milliseconds(10));
-            accepted.close();
-        } catch (...) {}
-    });
-
-    try {
-        Client<Http_client_connection> client(server_addr);
-        client.set_timeout(5);
-        client.execute("echo", Value("test"));
-        BOOST_FAIL("Expected exception for closed connection");
-    } catch (const iqnet::network_error& e) {
-        // Expected - connection closed
-        BOOST_CHECK(std::string(e.what()).find("closed") != std::string::npos ||
-                    std::string(e.what()).find("peer") != std::string::npos ||
-                    true);  // Any network error is acceptable
-    } catch (const std::exception&) {
-        // Other exceptions also acceptable
-        BOOST_CHECK(true);
-    }
-
-    acceptor.join();
-    server_sock.close();
-}
-
-// Test SSL context creation with invalid certificate path
-// Covers ssl_lib.cc lines 213-218: Certificate loading failure
-// Note: This test verifies that invalid cert paths are detected.
-// The ssl::exception() constructor has a known issue with NULL error strings,
-// so we test with a valid cert but invalid key to trigger a different code path.
-BOOST_AUTO_TEST_CASE(ssl_invalid_cert_path)
-{
-    iqnet::ssl::Ctx* saved_ctx = iqnet::ssl::ctx;
-
-    // Create a temp file that exists but is empty (invalid cert format)
-    std::string invalid_cert_path = "/tmp/iqxmlrpc_empty_cert.pem";
-    std::string invalid_key_path = "/tmp/iqxmlrpc_empty_key.pem";
-
-    std::ofstream cert_file(invalid_cert_path);
-    cert_file << "NOT A VALID CERTIFICATE";
-    cert_file.close();
-
-    std::ofstream key_file(invalid_key_path);
-    key_file << "NOT A VALID KEY";
-    key_file.close();
-
-    bool exception_thrown = false;
-    try {
-        // Try to create context with invalid certificate content
-        iqnet::ssl::Ctx* ctx = iqnet::ssl::Ctx::client_server(
-            invalid_cert_path,
-            invalid_key_path);
-
-        // Should not reach here
-        delete ctx;
-    } catch (...) {
-        // Any exception is expected - invalid cert format
-        exception_thrown = true;
-    }
-
-    std::remove(invalid_cert_path.c_str());
-    std::remove(invalid_key_path.c_str());
-    iqnet::ssl::ctx = saved_ctx;
-
-    BOOST_CHECK(exception_thrown);
-}
-
-// Test SSL context with mismatched cert/key
-// Covers ssl_lib.cc line 216: SSL_CTX_check_private_key failure
-BOOST_AUTO_TEST_CASE(ssl_mismatched_cert_key)
-{
-    iqnet::ssl::Ctx* saved_ctx = iqnet::ssl::ctx;
-
-    // Create a cert file with different content than key
-    std::string cert_path = "/tmp/iqxmlrpc_test_mismatch_cert.pem";
-    std::string key_path = "/tmp/iqxmlrpc_test_mismatch_key.pem";
-
-    // Write a valid cert
-    std::ofstream cert_file(cert_path);
-    cert_file << EMBEDDED_TEST_CERT;
-    cert_file.close();
-
-    // Write an invalid key (this will fail during key loading)
-    std::ofstream key_file(key_path);
-    key_file << "-----BEGIN PRIVATE KEY-----\nINVALIDKEY\n-----END PRIVATE KEY-----\n";
-    key_file.close();
-
-    bool exception_thrown = false;
-    try {
-        iqnet::ssl::Ctx* ctx = iqnet::ssl::Ctx::client_server(cert_path, key_path);
-        delete ctx;
-    } catch (...) {
-        // Any exception is expected - key invalid or doesn't match cert
-        exception_thrown = true;
-    }
-
-    std::remove(cert_path.c_str());
-    std::remove(key_path.c_str());
-    iqnet::ssl::ctx = saved_ctx;
-
-    BOOST_CHECK(exception_thrown);
-}
-
-// Test ssl::exception with specific error code
-// Covers ssl_lib.cc lines 280-287: exception constructor with error code
-BOOST_AUTO_TEST_CASE(ssl_exception_with_error_code)
-{
-    // Create exception with a known error code
-    iqnet::ssl::exception ex(0x12345678);  // Arbitrary error code
-    BOOST_CHECK(std::string(ex.what()).find("SSL") != std::string::npos);
-}
-
-// Test ssl::io_error exception
-// Covers ssl_lib.cc lines 328-329: io_error exception
-BOOST_AUTO_TEST_CASE(ssl_io_error_exception)
-{
-    iqnet::ssl::io_error err(42);  // Arbitrary error code
-    BOOST_CHECK(std::string(err.what()).find("42") != std::string::npos ||
-                err.what() != nullptr);  // Just check it's valid
-}
-
-// Test verifier that throws exception
-// Covers ssl_lib.cc lines 159-161: Exception handling in verify()
-namespace {
-class ThrowingVerifier : public iqnet::ssl::ConnectionVerifier {
-    int do_verify(bool, X509_STORE_CTX*) const override {
-        throw std::runtime_error("Test exception from verifier");
-    }
-};
-}
-
-BOOST_AUTO_TEST_CASE(ssl_verifier_exception_handling)
-{
-    // The verify() method catches exceptions and returns 0
-    ThrowingVerifier verifier;
-
-    // We can't directly test verify() without a real SSL connection,
-    // but we can verify the exception types exist and work
-    BOOST_CHECK_NO_THROW({
-        iqnet::ssl::need_read nr;
-        iqnet::ssl::need_write nw;
-        (void)nr.what();
-        (void)nw.what();
-    });
-}
-
-// Test HTTP proxy client URI decoration
-// Covers http_client.cc lines 76-87: decorate_uri edge cases
-BOOST_AUTO_TEST_CASE(http_proxy_uri_decoration)
-{
-    // Create a mock connection to test URI decoration
-    // Since Http_proxy_client_connection is protected, we test indirectly
-    // by checking that the client can be constructed
-
-    // This test primarily ensures the proxy client code paths are exercised
-    // The actual URI decoration happens when making requests
-
-    // Test with different URI formats would require a proxy server
-    // For now, just verify the class can be instantiated
-    Inet_addr addr("127.0.0.1", 8080);
-
-    // Proxy client requires actual proxy, so we just test construction doesn't crash
-    BOOST_CHECK_NO_THROW({
-        // The proxy connector is used with Https_proxy_client_connection
-        // We can't fully test without a proxy, but verify types compile
-        (void)addr.get_port();
-    });
-}
-
-BOOST_AUTO_TEST_SUITE_END()
-
-//=============================================================================
-// Reactor Error Path Tests - covers reactor_impl.h, reactor_poll_impl.cc
-//
-// These tests exercise error handling in reactor operations:
-//   - No handlers exception (reactor_impl.h lines 287-288)
-//   - Handler exception logging (reactor_impl.h lines 193-205)
-//=============================================================================
-
-BOOST_AUTO_TEST_SUITE(reactor_error_tests)
-
-// Test reactor with no active handlers
-// Covers reactor_impl.h lines 287-288: No_handlers exception
-BOOST_AUTO_TEST_CASE(reactor_no_handlers)
-{
-    // The No_handlers exception is thrown when reactor has no handlers
-    // We can test this by checking the exception type exists
-    BOOST_CHECK_NO_THROW({
-        try {
-            throw iqnet::Reactor_base::No_handlers();
-        } catch (const std::exception& e) {
-            BOOST_CHECK(std::string(e.what()).find("handler") != std::string::npos ||
-                        e.what() != nullptr);
-        }
-    });
-}
-
-// Test that server properly handles exceptions in methods
-// Covers reactor_impl.h lines 193-205: invoke_servers_handler exception handling
-BOOST_FIXTURE_TEST_CASE(server_method_exception_handling, IntegrationFixture)
-{
-    start_server(1, 130);
-
-    std::unique_ptr<Client_base> client(
-        new Client<Http_client_connection>(Inet_addr("127.0.0.1", port_)));
-
-    // Call method that throws std::exception
-    Response r1 = client->execute("std_exception_method", Param_list());
-    BOOST_CHECK(r1.is_fault());
-
-    // Call method that throws unknown exception type
-    Response r2 = client->execute("unknown_exception_method", Param_list());
-    BOOST_CHECK(r2.is_fault());
-
-    // Server should still be running after handling exceptions
-    Response r3 = client->execute("echo", Value("still alive"));
-    BOOST_CHECK(!r3.is_fault());
-    BOOST_CHECK_EQUAL(r3.value().get_string(), "still alive");
-}
-
-// Test socket operations error paths
-// Covers socket.cc various error conditions
-BOOST_AUTO_TEST_CASE(socket_error_operations)
-{
-    Socket sock;
-
-    // Test get_last_error on fresh socket
-    int err = sock.get_last_error();
-    BOOST_CHECK_EQUAL(err, 0);
-
-    // Test shutdown on unconnected socket (should not throw)
-    BOOST_CHECK_NO_THROW(sock.shutdown());
-
-    sock.close();
-}
-
-// Test Inet_addr edge cases
-BOOST_AUTO_TEST_CASE(inet_addr_operations)
-{
-    // Test with hostname
-    Inet_addr addr1("localhost", 8080);
-    BOOST_CHECK_EQUAL(addr1.get_port(), 8080);
-
-    // Test with IP address
-    Inet_addr addr2("127.0.0.1", 9090);
-    BOOST_CHECK_EQUAL(addr2.get_port(), 9090);
-    BOOST_CHECK_EQUAL(addr2.get_host_name(), "127.0.0.1");
-
-    // Test copy construction
-    Inet_addr addr3(addr2);
-    BOOST_CHECK_EQUAL(addr3.get_port(), addr2.get_port());
-}
-
-BOOST_AUTO_TEST_SUITE_END()
-
-//=============================================================================
-// Reactor Mask Tests
-//
-// These tests validate the reactor's event mask handling, specifically
-// the bitwise mask clearing operations (using ~mask instead of !mask).
-// Bug fix verified: reactor_impl.h lines 141 and 243.
-//=============================================================================
-
-BOOST_AUTO_TEST_SUITE(reactor_mask_tests)
-
-// Test that unregistering specific event masks works correctly
-// This validates the !mask -> ~mask fix by exercising mask transitions
-BOOST_FIXTURE_TEST_CASE(unregister_specific_event_mask, IntegrationFixture)
-{
-  start_server(1, 50);
-
-  // Verify server can handle multiple sequential requests
-  // This exercises the mask registration/unregistration path:
-  // - INPUT mask registered when waiting for request
-  // - OUTPUT mask registered when sending response
-  // - Masks cleared correctly between requests
-  auto client = create_client();
-
-  // First request - registers INPUT handler, then OUTPUT for response
-  Response r1 = client->execute("echo", Value("test1"));
-  BOOST_CHECK(!r1.is_fault());
-  BOOST_CHECK_EQUAL(r1.value().get_string(), "test1");
-
-  // Second request - same pattern, validates mask cleared correctly
-  Response r2 = client->execute("echo", Value("test2"));
-  BOOST_CHECK(!r2.is_fault());
-  BOOST_CHECK_EQUAL(r2.value().get_string(), "test2");
-
-  // Third request with keep-alive - exercises mask transitions
-  // INPUT -> OUTPUT -> INPUT again on same connection
-  client->set_keep_alive(true);
-  Response r3 = client->execute("echo", Value("test3"));
-  BOOST_CHECK(!r3.is_fault());
-  BOOST_CHECK_EQUAL(r3.value().get_string(), "test3");
-
-  Response r4 = client->execute("echo", Value("test4"));
-  BOOST_CHECK(!r4.is_fault());
-  BOOST_CHECK_EQUAL(r4.value().get_string(), "test4");
-}
-
-// Test that the reactor handles concurrent mask operations with pool executor
-// Pool executor uses mutex-protected reactor (std::mutex vs Null_lock)
-BOOST_FIXTURE_TEST_CASE(reactor_mask_with_pool_executor, IntegrationFixture)
-{
-  start_server(4, 51);  // 4 threads, unique port
-
-  // Make several concurrent requests to stress test mask handling
-  std::vector<std::thread> threads;
-  std::atomic<int> success_count(0);
-
-  for (int i = 0; i < 10; ++i) {
-    threads.emplace_back([this, i, &success_count]() {
-      try {
-        auto client = create_client();
-        Response r = client->execute("echo", Value(i));
-        if (!r.is_fault() && r.value().get_int() == i) {
-          ++success_count;
-        }
-      } catch (...) {}
-    });
-  }
-
-  for (auto& t : threads) {
-    t.join();
-  }
-
-  // All requests should succeed
-  BOOST_CHECK_EQUAL(success_count.load(), 10);
-}
-
-// Test mask transitions with various data types
-// Exercises different code paths through the reactor with different payload sizes
-BOOST_FIXTURE_TEST_CASE(reactor_mask_complex_responses, IntegrationFixture)
-{
-  start_server(1, 52);
-  auto client = create_client();
-
-  // Request that returns a struct (different serialization path)
-  Struct s;
-  s.insert("value", Value(42));
-  s.insert("name", Value("test"));
-  Response r1 = client->execute("echo", Value(s));
-  BOOST_CHECK(!r1.is_fault());
-  BOOST_CHECK_EQUAL(r1.value()["value"].get_int(), 42);
-
-  // Request that returns an array
-  Array arr;
-  arr.push_back(Value(1));
-  arr.push_back(Value(2));
-  arr.push_back(Value(3));
-  Response r2 = client->execute("echo", Value(arr));
-  BOOST_CHECK(!r2.is_fault());
-  BOOST_CHECK_EQUAL(r2.value().size(), 3u);
-}
-
-BOOST_AUTO_TEST_SUITE_END()
-
-//=============================================================================
-// Defensive Synchronization Tests
-//
-// These tests exercise the defensive mutex protections added to:
-//   - Server connection set (server.cc)
-//   - Idle connection state (server_conn.cc)
-//   - SSL partial write handling (ssl_connection.cc)
-//=============================================================================
-
-BOOST_AUTO_TEST_SUITE(defensive_sync_tests)
-
-// Test that concurrent client requests work correctly with connection mutex
-// Exercises server.cc connection set synchronization
-BOOST_FIXTURE_TEST_CASE(concurrent_connection_registration, IntegrationFixture)
-{
-  start_server(4, 60);  // Pool executor with 4 threads
-
-  // Launch multiple concurrent clients
-  std::vector<std::thread> threads;
-  std::atomic<int> success_count(0);
-
-  for (int i = 0; i < 20; ++i) {
-    threads.emplace_back([this, i, &success_count]() {
-      try {
-        auto client = create_client();
-        Response r = client->execute("echo", Value(i));
-        if (!r.is_fault() && r.value().get_int() == i) {
-          ++success_count;
-        }
-      } catch (...) {}
-    });
-  }
-
-  for (auto& t : threads) {
-    t.join();
-  }
-
-  BOOST_CHECK_EQUAL(success_count.load(), 20);
-}
-
-// Test large data transfer (exercises SSL partial write loop)
-// Covers ssl_connection.cc send() retry loop for partial writes
-BOOST_FIXTURE_TEST_CASE(large_data_transfer, IntegrationFixture)
-{
-  start_server(1, 61);
-  auto client = create_client();
-
-  // Create a large string to transfer (64KB)
-  std::string large_data(65536, 'X');
-  Response r = client->execute("echo", Value(large_data));
-
-  BOOST_CHECK(!r.is_fault());
-  BOOST_CHECK_EQUAL(r.value().get_string().length(), large_data.length());
-}
-
-// Test rapid connection create/destroy cycles
-// Exercises connection registration/unregistration under load
-BOOST_FIXTURE_TEST_CASE(rapid_connection_cycling, IntegrationFixture)
-{
-  start_server(2, 62);
-
-  for (int cycle = 0; cycle < 10; ++cycle) {
-    std::vector<std::thread> threads;
-    std::atomic<int> success_count(0);
-
-    // Create burst of connections
-    for (int i = 0; i < 5; ++i) {
-      threads.emplace_back([this, &success_count]() {
-        try {
-          auto client = create_client();
-          Response r = client->execute("echo", Value("rapid"));
-          if (!r.is_fault()) {
-            ++success_count;
-          }
-        } catch (...) {}
-      });
-    }
-
-    for (auto& t : threads) {
-      t.join();
-    }
-
-    BOOST_CHECK_GE(success_count.load(), 3);  // At least 60% success
-  }
-}
-
-// Test idle state transitions with keep-alive
-// Exercises idle_mutex_ in server_conn.cc
-BOOST_FIXTURE_TEST_CASE(idle_state_transitions, IntegrationFixture)
-{
-  start_server(1, 63);
-
-  auto client = create_client();
-  client->set_keep_alive(true);
-
-  // Make multiple requests on same connection
-  // Each request: stop_idle -> process -> start_idle
-  for (int i = 0; i < 5; ++i) {
-    Response r = client->execute("echo", Value(i));
-    BOOST_CHECK(!r.is_fault());
-    BOOST_CHECK_EQUAL(r.value().get_int(), i);
-  }
-}
-
-BOOST_AUTO_TEST_SUITE_END()
+// NOTE: defensive_sync_tests moved to test_integration_advanced.cc
 
 //=============================================================================
 // Coverage Improvement Tests - ssl_lib.cc, https_server.cc, http_client.cc
